Switch to new API, edit exposed entities
As the old API app.greenchoice.nl seems to have been taken offline, I have replicated the sign in process from a web browser.
This method requires various requests to be sent to mijn.greenchoice.nl and also sso.greenchoice.nl

I have no experience with API's whatsoever, so this might not be the best approach, but it works for now.

Also, it seems easier to expose the total energy consumption of the current year, and use the energy statistics dashboard in homeassistant to keep track of daily, monthly usage.
I have also exposed four price entities: costEnergyKwh, costEnergyDailyBase, costGasM3, costGasDailyBase. To keep track of the price per unit, but also the daily base price that is paid.